Updated JWT tests in JDBC and client
