+++ client-v2/src/test/java/com/clickhouse/client/HttpTransportTests.java
@@ -1563,33 +1563,6 @@ public class HttpTransportTests extends BaseIntegrationTest {
     }
 
     @Test(groups = { "integration" })
-    public void testJWTWithCloud() throws Exception {
-        if (!isCloud()) {
-            return; // only for cloud
-        }
-        final String jwt = System.getenv("CLIENT_JWT");
-        final String host = System.getenv("JWT_TEST_HOST");
-        Assert.assertTrue(jwt != null && !jwt.trim().isEmpty(), "CLIENT_JWT is not set.");
-        Assert.assertTrue(host != null && !host.trim().isEmpty(), "JWT_TEST_HOST is not set");
-        Assert.assertFalse(jwt.contains("\n") || jwt.contains("-----"), "JWT should be single string ready for HTTP header");
-        try (Client client = new Client.Builder()
-                .addEndpoint(Protocol.HTTP, host, 8443, true)
-                .setUsername("default")
-                .compressClientRequest(false)
-                .setDefaultDatabase("default")
-                .serverSetting(ServerSettings.WAIT_END_OF_QUERY, "1")
-                .useBearerTokenAuth(jwt).build()) {
-            try {
-                List<GenericRecord> response = client.queryAll("SELECT user(), now()");
-                System.out.println("response: " + response.get(0).getString(1) + " time: " + response.get(0).getString(2));
-            } catch (Exception e) {
-                e.printStackTrace();
-                throw e;
-            }
-        }
-    }
-
-    @Test(groups = { "integration" })
     public void testWithDefaultTimeouts() {
         if (isCloud()) {
             return; // mocked server
